add loggers

+++ InDBClient.py
@@ -4,13 +4,19 @@ import argparse
 import threading
 import time
 import sys
+import logging
 
-import pyximport; pyximport.install()
+from argparse import RawTextHelpFormatter
+import pyximport
+from scapy.plist import QueryAnswer; pyximport.install()
 import InDBCollector
 
-def parse_params():
-    parser = argparse.ArgumentParser(description='InfluxBD INTCollector client.')
+logging.basicConfig(level = logging.INFO)
+logger = logging.getLogger(__name__)
 
+def parse_params():
+    parser = argparse.ArgumentParser(description='InfluxBD INTCollector client.')#, formatter_class=RawTextHelpFormatter)
+    
     parser.add_argument("ifaces", nargs='+',
     help="List of ifaces to receive INT reports")
 
@@ -26,7 +32,7 @@ def parse_params():
     parser.add_argument("-p", "--period", default=10, type=int,
         help="Time period to push data in normal condition")
 
-    parser.add_argument("-P", "--event_period", default=1, type=float,
+    parser.add_argument("-P", "--event_period", default=0, type=float,
         help="Time period to push event data")
 
     parser.add_argument("-t", "--int_time", action='store_true',
@@ -34,10 +40,18 @@ def parse_params():
 
     parser.add_argument("-e", "--event_mode", default="THRESHOLD",
         help="Event detection mode: INTERVAL or THRESHOLD. \
-        Option -p is disabled for THRESHOLD and is hard-coded instead")
+            Option -p is disabled for THRESHOLD and is hard-coded instead")
 
-    parser.add_argument("-d", "--debug_mode", default=0, type=int,
-        help="Set to 1 to print event")
+    parser.add_argument("-l", "--log_level", default=20, type=int,
+        help="CRITICAL = 50\
+            ERROR = 40;\
+            WARNING = 30;\
+            INFO = 20;\
+            DEBUG = 10;\
+            NOTSET = 0;")
+    
+    parser.add_argument("-l_rap", "--log_raports_lvl", default=20, type=int,
+        help='DEBUG = 10 - enables logging of raports')
 
     return parser.parse_args()
 
@@ -46,50 +60,49 @@ if __name__ == "__main__":
 
     args = parse_params()
 
-    # number_of_events = 0
-
-    collector = InDBCollector.InDBCollector(int_dst_port=args.int_port,
-        debug_mode=args.debug_mode, host=args.host,
-        database=args.database, int_time=args.int_time,
-        event_mode=args.event_mode)
+    logger.setLevel(args.log_level)
+    
+    collector = InDBCollector.InDBCollector(int_dst_port=args.int_port, 
+        host=args.host, database=args.database, 
+        int_time=args.int_time, event_mode=args.event_mode, 
+        log_level=args.log_level, log_raports_lvl = args.log_raports_lvl
+        )
 
 
     for iface in args.ifaces:
         collector.attach_iface(iface)
 
     # clear all old dbs. For easy testing
-    for db in collector.client.get_list_database():
-        collector.client.drop_database(db["name"])
-    collector.client.create_database(args.database)
+    clear_db: str = 'n'
+    clear_db = input(f'Database name: {args.database}.\nShould I clear the database? [y/n]: ')
+    if clear_db in ['yes', 'y', 'Y', 'YES']:
+        for db in collector.client.get_list_database():
+            collector.client.drop_database(db["name"])
+        collector.client.create_database(args.database)
+        logger.info(f'Database has been cleared.')
 
+    databases_list = [x['name'] for x in collector.client.get_list_database()]
+    if not(args.database in databases_list):
+        collector.client.create_database(args.database)
+        logger.info(f'Databases created.')
 
     push_stop_flag = threading.Event()
 
-    # A separated thread to push event data
+    # # A separated thread to push event data
     def _event_push():
 
         while not push_stop_flag.is_set():
 
-            # time.sleep(args.event_period)
-
+            time.sleep(args.event_period)
+ 
             collector.lock.acquire()
             data = collector.event_data
             collector.event_data = []
             collector.lock.release()
 
-            if args.debug_mode==2:
-                print("Len of events: ", len(data))
-                # number_of_events = number_of_events+ 1
-                # print(f'{number_of_events}')                
-
             if data:
                 collector.client.write_points(points=data[0])
-                # print(type(data))
-                # print(data)
-                # for x in range(len(data[0])):
-                    # collector.client.write_points(points=data[0][x])
-                    # points=data[0][x]
-                    # print(type(points))
+                logger.debug(f'Len of data: {len(data)}')
 
 
     # A separated thread to push data
@@ -108,9 +121,7 @@ if __name__ == "__main__":
                 data = collector.collect_data()
                 if data:
                     collector.client.write_points(points=data, protocol=protocol)
-                    if args.debug_mode==2:
-                        print("Periodically push: ", len(data))
-
+                    logger.debug(f'Periodically push: {len(data[0])}')
 
         periodically_push = threading.Thread(target=_periodically_push)
         periodically_push.start()
